requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
colorlog
pyjwt
passlib[bcrypt]
email-validator